# openapi/schema.yaml
openapi: "3.0.3"
info:
  title: Do Later API
  description: Do Later API
  version: 1.0.0
servers:
  - url: https://api.dolater.kantacky.com
    description: Production Server
  - url: https://stg.api.dolater.kantacky.com
    description: Staging Server
  - url: https://dev.api.dolater.kantacky.com
    description: Development Server
  - url: http://localhost:8080
    description: Local Server
paths:
...
  /users/{uid}/follow:
    get:
      tags:
        - user
      summary: Get Follow Reqest
      description: Get follow request from a user
      operationId: getFollowRequest
      security:
        - AppCheck: []
        - AuthBearer: []
      parameters:
        - $ref: "#/components/parameters/uid"
      responses:
        200:
          $ref: "#/components/responses/FollowStatusOK"
        401:
          $ref: "#/components/responses/Unauthorized"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - user
      summary: Follow User
      description: Request to follow a user
      operationId: followUser
      security:
        - AppCheck: []
        - AuthBearer: []
      parameters:
        - $ref: "#/components/parameters/uid"
      responses:
        200:
          $ref: "#/components/responses/FollowStatusOK"
        401:
          $ref: "#/components/responses/Unauthorized"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/InternalServerError"
    patch:
      tags:
        - user
      summary: Approve Follow Request
      description: Approve follow request from a user
      operationId: approveFollowRequest
      security:
        - AppCheck: []
        - AuthBearer: []
      parameters:
        - $ref: "#/components/parameters/uid"
      responses:
        200:
          $ref: "#/components/responses/FollowStatusOK"
        401:
          $ref: "#/components/responses/Unauthorized"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags:
        - user
      summary: Unfollow User
      description: Unfollow a user
      operationId: unfollowUser
      security:
        - AppCheck: []
        - AuthBearer: []
      parameters:
        - $ref: "#/components/parameters/uid"
      responses:
        204:
          $ref: "#/components/responses/NoContent"
        401:
          $ref: "#/components/responses/Unauthorized"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/InternalServerError"
...
  /tasks:
    get:
      tags:
        - task
      summary: Get Tasks
      description: Get all tasks
      operationId: getTasks
      security:
        - AppCheck: []
        - AuthBearer: []
      parameters:
        - $ref: "#/components/parameters/poolId"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
        401:
          $ref: "#/components/responses/Unauthorized"
        500:
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - task
      summary: Create Task
      description: Create a task
      operationId: createTask
      security:
        - AppCheck: []
        - AuthBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskInput"
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        401:
          $ref: "#/components/responses/Unauthorized"
        500:
          $ref: "#/components/responses/InternalServerError"
...
  /notifications:
    get:
      tags:
        - notification
      summary: Get Notifications
      description: Get notifications
      operationId: getNotifications
      security:
        - AppCheck: []
        - AuthBearer: []
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
        401:
          $ref: "#/components/responses/Unauthorized"
        500:
          $ref: "#/components/responses/InternalServerError"
...
components:
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
        displayName:
          type: string
        photoURL:
          type: string
          format: uri
        activeTaskPool:
          $ref: "#/components/schemas/TaskPool"
        archivedTaskPool:
          $ref: "#/components/schemas/TaskPool"
        pendingTaskPool:
          $ref: "#/components/schemas/TaskPool"
        followings:
          type: array
          items:
            $ref: "#/components/schemas/FollowStatus"
        followers:
          type: array
          items:
            $ref: "#/components/schemas/FollowStatus"
      required:
        - id
        - displayName
        - photoURL
        - activeTaskPool
        - archivedTaskPool
        - pendingTaskPool
        - followings
        - followers
...
    FollowStatus:
      type: object
      properties:
        from:
          $ref: "#/components/schemas/User"
        to:
          $ref: "#/components/schemas/User"
        requestedAt:
          type: string
          format: date-time
        approvedAt:
          type: string
          format: date-time
      required:
        - from
        - to
        - requestedAt
    TaskPool:
      type: object
      properties:
        id:
          type: string
          format: uuid
        owner:
          $ref: "#/components/schemas/User"
      required:
        - id
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        archivedAt:
          type: string
          format: date-time
        owner:
          $ref: "#/components/schemas/User"
        pool:
          $ref: "#/components/schemas/TaskPool"
      required:
        - id
        - url
        - createdAt
        - owner
        - pool
    CreateTaskInput:
      type: object
      properties:
        url:
          type: string
          format: uri
      required:
        - url
...
    Notification:
      type: object
      properties:
        title:
          type: string
        body:
          type: string
        url:
          type: string
          format: uri
        createdAt:
          type: string
          format: date-time
      required:
        - title
        - createdAt
  parameters:
...
    uid:
      name: uid
      in: path
      required: true
      description: UID
      schema:
        type: string
    poolId:
      name: poolId
      in: query
      required: false
      description: Pool ID
      schema:
        type: string
        format: uuid
  securitySchemes:
    AuthBearer:
      type: http
      scheme: bearer
      description: Bearer Access Token for Authorization
    AppCheck:
      type: apiKey
      in: header
      name: X-Firebase-AppCheck
      description: Firebase App Check Token
...
  responses:
...
    FollowStatusOK:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/FollowStatus"
...
    NoContent:
      description: No Content
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            message: Unauthorized
    NotFound:
      description: NotFound
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            message: NotFound
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            message: Internal Server Error